Provide a simple tacacs communication library for TACACS+ accounting
Initial release of simple tacacs accounting library so other shared
libraries can connect to a TACACS+ server without being tacacs aware.

This library links against libtac.so, normally provided by the
pam-tacplus package

It provides a dev package as well as a runtime package.  It is tested
on debian wheezy and debian jessie